Removing section on database TDD.
I think TDD for a database is fine - but not for its schema.  Writing
tests for a schema (tables, their columns, etc.) seems pointless, since
you will write what you want it to look like in the database versioning
scripts, and you can see the current state of the database easily enough
in database viewing applications.

On the other hand, for the purposes of testing delete cascades,
triggers, procedures, constraints, and other logic, I would say go ahead
and write the unit tests.

How to do so is an exercise I will leave for the reader.