Adding rudimentary database access layer and tests.
Our test doesn't pass yet - but it will soon.  Right now we are
failing because every run of the test adds a new row to our "person"
table, but we want the ability to dump and restore, so we can set
the database into whatever state we want instantly.  We'll do that next.

Currently using Postgresql 10.6