Refactor and correction.
Moving the values for the restore command and the path of database backup files to constants.  Perhaps these should be set as configurable items, or need to be called out in the developer docs.

Also, added a correction to the code that runs the restore command - before, it was kicking off the command and not waiting until it completed.  I changed this so it now waits.