cascade of changes because of using real database and moc for borrower.
Put some constants for universal things like the URL to the database in
the Constants file.  Put the names of particular database backup
files in the database_backup_constants so we have documentation and
don't have to go looking in the directories guessing at what each file
holds.

This is just the first step.  Ultimate goal is to use the persistence
layer or mocks of it for all things that are currently being done using
simple files as databases.