Refactoring, adding comments, adding assertions in non-test code.
assertions are meant to represent inviolable things within some logic.
They are not as much for purposes of checking inputs to a function (where
you might want to pass along a message to the caller about why things
failed) as they are for indicating that something simply impossible
has happened, and to illuminate it. That is, it's more to indicate that
something complex hasn't been fully coded correctly.

I added some checks at the beginning of methods to make sure that
certain things are true, like that we are given valid arguments.  I also
extracted those methods into their own helper methods and added a class
to contain them.