Big refactor.  adding connection pools.
I found that when running the Zap OWASP proxy against my system, I was
running out of connections.  I realized I was leaking connections
profusely, and I wanted to use connection pooling (look that up).

By using a datasource, and using that to *get* a connection and then
close it when we're finished, I am able to have my cake and eat it.

At least, that's the assumption.  Let's see how it works.