Choosing a non-memory location for h2 when running tests.
Found that it was difficult to observe what was happening with
the unit tests, because I wasn't able to see the same database
after the unit tests finished.  By setting it to a physical
file, it lasts a bit longer and I can use

gradlew startConsole

to see what's in it and step through.