Using a Future for running Nbvcxz, and using SLF4j logger.
I found that when running Zap against the application, it seemed
like I was regularly encountering issues with Nbvcxz getting hung
up working while Tomcat was trying to close things down.  For that
reason it seemed like this is the one thing in the system that has a
chance of taking too long, which means - performance problem.

I wrapped it in a Future, so when a whole second has passed, it times
out the work and returns a result indicating that, and logs a warning.