Reverting to non-promise usage of Nbvcxz
Well, I had hoped to mitigate my perf problems by wrapping the
estimation tool in a future (basically, a thread) that I could control.
But it turned out that wasn't as easy as I hoped, and a far simpler
solution presented itself when I took a good hard look at the system
I was running on - it only had 100 megabytes of free memory.  When I ran
an experiment with Nbvcxz, I found it often uses 1.5 gigs while running.
ok.  I upped the memory on the test box.  Let's see if that solves all.