Adjusting to use Chromedriver in path or by setting its path.
Originally we were able to use Chromedriver in one of two cases:
Either it was located in a directory that was in the PATH, or..
it was in the ui_tests directory.

However, that second option did not always work, depending on which
directory you were in when you ran the program.

A better way.  Now, we check if it's on the PATH or it's set in
gradle.properties.

Also, improving the error message if Chromedriver cannot be found.