gradle now adds a pre-push git hook at clean.
Alright, let me break this down for you.

What we really want to happen is that we don't want to allow
a user to push to the central repo unless unit tests pass.  That
takes very little time and does a good job of a extremely
preliminary smoke test.

In order to automate the process of creating this git hook, we
added a script to gradle that will move the file to the proper location
whenever a git clean runs.

If all the unit tests (and *only* the unit tests) pass, only then
we will send the data to the central repo.  Oh, one other thing...

We do a git stash before running tests.  So we'll only be running
tests on code that's been fully committed.