Handling deletion of books / borrowers better.
if a book or a borrower is deleted from the library, what happens?

Well, truth is, depends.  Perhaps in more sophisticated systems, we
never actually delete anything (so we keep all data relationships
clear) but we set data like books or borrowers as "inactive".

However, we aren't worried in MVP-world with such sophisticated niceties.
We just want the simplest thing that works.  Therefore, we really
delete - and what we're deciding to do here is to delete any loans
that are associated with the books or borrowers when they are removed.