Dealing with scenarios: deleting books/borrowers, multiple loans.
If a book or a borrower is deleted, what happens?  In this commit, we
added some scenarios in relevant features to describe what should occur.

Also, it came up - have we been clear that a borrower can borrow
multiple books?  And, how do we see which books a borrower has?