Fixing problem preventing generateCucumberReport.
Secondarily, updating dependency check to the latest version,
updating gradle-cucumber-report to the latest version, causing
a change to its configuration.

Details:
  both gradle-cucumber-reporting and dependencycheck use
apache.velocity.  Problem is that dependencycheck uses the most recent
version of it, 2.1.  When cucumberreport tries to run, gradle gives it,
per its default, the most recent version - 2.1, which causes it to
crash.

I needed to do something so that these two plugins will go looking in
different places to find their dependencies.  I was initially planning
to do some work with the "exclude" command in the classpath in the
buildscript, but after using that syntax and testing it out, I didn't
even need to go that far.

Apparently simply having the dependencycheck plugin listed in "plugins"
and the gradle-cucumber-reporting in "buildscript" was sufficient to
cause the dependencies to no longer conflict.  Yay!