pre-push is now a document, doesn't get auto-copied
before this commit, I had the system putting the pre-push
script into the git directory whenever we ran clean.  However, one
of the commands to do so was chmod, which doesn't exist on many
Windows machines, so it would fail at that step.

Fixing it so that would work on Windows machines occurred to me, but
I decided that I would rather have fewer moving parts, so I will
simply have the pre-push script located in the /doc directory if
anyone wishes to use it, and won't automatically install it.