Our proof of concept works!
I was able to control the UI through the client application!

When you start the UI now, it starts a separate thread running the
server.  When you use the client program to connect to the server,
anything you type gets shown in the bottom of the UI window. So we
have proven we can do what we want.  MVP complete!

Next steps for features / bugs:

1) Simplify the UI down and make it work for the autoinsurance domain.
2) Provide more commands so that a tester can automate everything
     needed to test the functioning
3) Try for fewer moving parts in general.