Hooked up UI and now and able to manipulate UI elements.
I added an action listener for the button to pull the data
from the age and claims fields and calculate the result, and
put that in the label field.

I added some code to programmatically manipulate the values
in the claims drop-down, the age field, and the calculate button.

So far so good.