Adding tail-recursive versions of Fibonacci and Ackermann
The recursive algorithms are simpler to read, but they are slower
and use more memory.  These new algorithms can go further with the
memory a computer has.  Much less wasteful.

Interesting note - the Ackermann's tail-recursive algorithm comes
up with an incorrect answer for M = 4 and N = 0

Added UI for these - select from a dropdown to choose the algorithm
you want to use.