Completing necessary scaffolding code for BDD to run
Adding a section in the build file to run the new expenses
and cartesian product features, and altering the cartesian
product text so it parses properly in Gherkin.

Everything runs now, though these two features show as
failing in the report, as they ought to, not having started
development yet.