Modifying the build file and new features have pending steps.
Made a change in the build file - I've divided up the cucumber task
into five separate tasks and created a new task to run them all.  This
enables me to bit a bit more organized, and also to print info as
we transition from one feature to the next.

I've added step definitions and some basic (empty) methods for the
features, as well as some initial ideas for data types.

Everything is working but the pending feature shows as such in the report.