Switching to using a "RESTful" output for APIs
When a user adds a book, they are sent to a "results" page to
see the result of their action.

For APIs like math (adding, Fibonacci, Ackermann) and listing
books and borrowers, I switched it to use a "RESTful" JSP page that
simply outputs the raw data, no HTML involved.

Making changes to the tests and SOAPui tests for this to work.