proof of concept for AWS and have ssh commands.
What commands do I need to provision and deploy to systems?
ssh - an ability to securely run commands on another machine DONE
scp - an ability to securely copy files to/from another machine DONE
aws - an ability to control AWS. (available as a library with groovy)  DONE.

If you want to see the test command to ssh: gradlew checkWebServers

if you want to create an instance with AWS, go to StartAws and run the unit test there.

Now the fun part, actually configuring all this stuff.