Simplifying the desktop UI quite a bit.
We have a capability in Demo to calculate auto insurance
premiums, based on prior numbers of claims and driver age.
This is provided through a Java Swing desktop UI interface.
We have additional code provided to enable programmatic
control of this UI.

We will get this more refined in the future, but for now I
am chopping out all the extra fluff that had nothing to do
with this need.