Trying a different (better?) tack on desktop tests
We'll use the sockets to connect and take actions, but we'll
drive it through JUnit, so we get some of that awesomeness.

I'm chopping some of the functionality in the name of simplicity.
Gone is the ability to send a list of files, each of which has
lines of script.

Next up: tie this in to everything else test-wise, get
jacoco test coverage working smoothly.