Extract AutoInsurance to its own project.
It was difficult to determine code coverage on the autoinsurance
desktop application, and I realized this was due to its inherently
different paradigm.  I therefore took the appropriate next step,
moving it to its own isolated project.

If you wish to interact with the autoinsurance desktop app, simply
change your directory into /autoinsurance and see the gradle
commands there.