moving style aspects to css and better handling of outside clicks.
Several aspects of the search box, like its width, are now taken from
the css file - specifically main.css

also, we want a situation where clicking outside the searchbox causes it
to go away, doing nothing.  This turned out to be trickier than I would
have thought.  Why?  here's why:
1) I could do onBlur, which fires whenever an element loses focus.  I had
that on the input field.  however when I tried clicking an element in the
list, it fired, so I could not select an element!
2) However, onblur is super valuable for when i tab away - that's really
the only way I currently know to close the searchbox if I was in the input
and then tabbed away, and I have to check to make sure I am hitting a
"relatedTarget"
3) I still need a way, without using a general onblur, to close the searchbox,
so I have a click event that checks some logic.