Expanding on Java UI tests, added HtmlUnit
The UI tests run really fast, but I wanted to see how the system handled when running without JavaScript.  To get two birds with one stone, I also wanted to try a a different browser on the tests to make sure nothing depends on Chrome in particular.  I found HtmlUnit for this task.

One annoyance - the API includes many methods that throw checked exceptions.  I mean ... this is a testing tool, why would I want to be forced to handle exceptions?  Wrapped those to avoid the annoyance.

Also put in a timer on the tests to show the speed difference... the HtmlUnit tests run really fast.