Including non-BDD UI tests
BDD is fantastic for capabilities, but isn't the right tool for testing that the UI works as expected in all combinations.  Plain ol' test drivers are best.

Here, we're using HtmlUnit and Selenium to test - HtmlUnit is a fast headless browser, and we've even got JavaScript turned off, to make sure our application works even without JS running (also known as graceful degradation).

We're using Selenium to test that the JavaScript components work as expected in a few forms.