Handling JSON escaping properly now
JSON disallows certain characters in string.  Stuff like double quote, and so on.  It is corrected by properly escaping the text.

That is, double quote becomes backslash doublequote

It was found, prior to this fix, that entering a book or borrower name
that included a double quote would cause the javascript UI code to fail,
because it could not properly parse the JSON text.

Correctly escaping characters so JSON will parse