Refactoring - renaming and small corrections.
Starting to use invariants - stating the assumptions about the state within the code.  This is in order to both document the code, as well as to provide guarantees of correctness.  If a guarantee is not met, it will fail "fast and loud".

Marking files as executable

Git didn't have some files stored as executable, so if someone were to download them, they would need to change the "mode" of the file to executable first before it would run.  By doing this, we are handling it for them, so the user doesn't need to.

invariants are simply things that must be true - I am adding code to assert that what *must* be true is true.  This does two nice things:
a) it documents my understanding
b) it sometimes finds bugs

Small corrections.

Ignoring the perf test because on Amazon EC2 it sometimes fails, because Amazon hands out slow computers pretty often.