Trying to set env vars in Jenkinsfile
This way the API tests go through the Zap proxy, but if I do it on the command line I don't need that to be running.

Previous to this, the API tests required the Zap proxy, or *some* proxy, to be running at 9888 on the localhost.  This way, that is only a concern while the pipeline is running.